python/address-book/pyproject.toml
[project]
name = "address-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ecdsa>=0.19.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
